--- a/app/javascript/mastodon/features/ui/components/bundle.js
+++ b/app/javascript/mastodon/features/ui/components/bundle.js
@@ -1,27 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
 
-import { fetchBundleRequest, fetchBundleSuccess, fetchBundleFail } from '../../../actions/bundles';
-
-const mapDispatchToProps = dispatch => ({
-  onFetch () {
-    dispatch(fetchBundleRequest());
-  },
-  onFetchSuccess () {
-    dispatch(fetchBundleSuccess());
-  },
-  onFetchFail () {
-    dispatch(fetchBundleFail());
-  },
-});
-
-// https://reacttraining.com/react-router/web/guides/code-splitting
 class Bundle extends React.Component {
 
   static propTypes = {
-    load: PropTypes.func.isRequired,
-    retry: PropTypes.func.isRequired,
+    fetchComponent: PropTypes.func.isRequired,
+    loading: PropTypes.func.isRequired,
+    error: PropTypes.func.isRequired,
     children: PropTypes.func.isRequired,
     onFetch: PropTypes.func.isRequired,
     onFetchSuccess: PropTypes.func.isRequired,
@@ -29,7 +14,7 @@ class Bundle extends React.Component {
   }
 
   state = {
-    mod: null,
+    mod: undefined,
   }
 
   componentWillMount() {
@@ -37,36 +22,43 @@ class Bundle extends React.Component {
   }
 
   componentWillReceiveProps(nextProps) {
-    if (nextProps.load !== this.props.load) {
+    if (nextProps.fetchComponent !== this.props.fetchComponent) {
       this.load(nextProps);
     }
   }
 
   load = (props) => {
-    const { load, onFetch, onFetchSuccess, onFetchFail } = props || this.props;
+    const { fetchComponent, onFetch, onFetchSuccess, onFetchFail } = props || this.props;
+
+    this.setState({ mod: undefined });
     onFetch();
 
-    return load()
+    return fetchComponent()
       .then((mod) => {
         this.setState({ mod: mod.default });
         onFetchSuccess();
       })
       .catch(() => {
-        this.setState({ mod: this.retry });
+        this.setState({ mod: null });
         onFetchFail();
       });
   }
 
-  retry = (props) => {
-    const { retry: Retry } = this.props;
-
-    return <Retry onLoad={this.load} {...props} />;
-  }
-
   render() {
-    return this.props.children(this.state.mod);
+    const { loading: Loading, error: Error, children } = this.props;
+    const { mod } = this.state;
+
+    if (mod === undefined) {
+      return <Loading />;
+    }
+
+    if (mod === null) {
+      return <Error onRetry={this.load} />;
+    }
+
+    return children(mod);
   }
 
 }
 
-export default connect(null, mapDispatchToProps)(Bundle);
+export default Bundle;