--- a/app/javascript/mastodon/features/ui/components/bundle.js
+++ b/app/javascript/mastodon/features/ui/components/bundle.js
@@ -43,7 +43,7 @@ class Bundle extends React.Component {
     }
   }
 
-  componentDidUpdate (prevProps) {
+  componentDidUpdate () {
     this.props.onRender();
   }