--- a/app/javascript/mastodon/features/ui/components/column_loading.js
+++ b/app/javascript/mastodon/features/ui/components/column_loading.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import Column from '../../../components/column';
+import ColumnHeader from '../../../components/column_header';
+
+const ColumnLoading = () => (
+  <Column>
+    <ColumnHeader icon=' ' title='' multiColumn={false} />
+    <div className='scrollable' />
+  </Column>
+);
+
+export default ColumnLoading;