… app/javascript/mastodon/features/ui/components/modal_loading.js
… app/javascript/mastodon/features/ui/components/modal_root.js
… app/javascript/mastodon/features/ui/containers/bundle_container.js
… app/javascript/mastodon/features/ui/index.js
… app/javascript/mastodon/features/ui/util/async-components.js
… app/javascript/mastodon/features/ui/util/react_router_helpers.js
… app/javascript/mastodon/reducers/compose.js
… app/javascript/mastodon/reducers/index.js
… app/javascript/mastodon/reducers/media_attachments.js
… app/javascript/mastodon/store/configureStore.js
… app/javascript/styles/components.scss
… app/views/layouts/application.html.haml