Roadmap:
- [x] Display progress bar while fetching bundles
- [x] Routes
   - [x] Lazy load ([react-router guide](https://reacttraining.com/react-router/web/guides/code-splitting))
   - [x] Loading state
      - [x] Empty column
      - [x] Move styles to SCSS
   - [x] Replay request (if bundle could not be fetched)
- [x] Columns:
  - [x] Compose (+ reducer)
  - [x] Home
  - [x] Notifications (+ reducer)
  - [x] Public timeline
  - [x] Community timeline
  - [x] Hashtag timeline
  - [x] Loading state - empty column
- [x] Modals
  - [x] Components:
    - [x] Media
    - [x] Onboarding
    - [x] Video
    - [x] Boost
    - [x] Confirm
    - [x] Report
  - [x] Replay request (if bundle could not be fetched)
  - [x] Move styles to SCSS
- [x] Do not show the waiting state if the bundle is fetched fast enough (`~200ms`)
- [x] Optimize bundles - perhaps tweak `minChunks`?
- [x] Measure size improvement

Bundles: https://gist.github.com/sorin-davidoi/07def6eb30df015fbfc8ced9fa36babd