@sorin-davidoi 2017-06-21
When a modal bundle could not be fetched, the following is displayed, allowing users to retry the download
![image](https://user-images.githubusercontent.com/2109702/27383045-849885da-5689-11e7-9aa2-b2957458f855.png)

Same thing with route bundles
![image](https://user-images.githubusercontent.com/2109702/27383084-ae5829ac-5689-11e7-93c1-6f195dcfa3dd.png)