@sorin-davidoi 2017-06-21
This reduces the size of `application.js` by 27% (673kB -> 485kb), while creating ~25 extra bundles. The progress bar is being used to show the download progress of each bundle. While the bundle is fetched, it displays a loading state in accordance with the content. If the fetch fails, an error component is shown, which provides a "refresh" button, allowing the user to replay the download.

<img src="https://user-images.githubusercontent.com/2109702/27402982-0257a570-56c9-11e7-843d-ae268eb431a8.png" width="300" />

<img src="https://user-images.githubusercontent.com/2109702/27403085-676acc30-56c9-11e7-8982-3e93a549b375.png" width="300" />