@nightpool 2017-06-22
I'm not sure I see the practical benefit of this—won't it just increase latency on mobile due to the extra round trip requests?

Is there a way to measure whether the decreased bundle size will actually reflect real world usage?