@sorin-davidoi 2017-06-22
As a side note, it would be nice to set up Lighthouse integration, to measure these metrics for each PR.