@sorin-davidoi 2017-06-22
Unscientific findings (running Lighthouse against current master and this branch):
 - Performance score: +3 (51 -> 54)
 - Render blocking scripts: -200ms (2070ms -> 1820ms)
 - First paint: -140ms (7060ms -> 6950ms)
 - First interactive: -600ms (7570ms -> 6950ms)

I would prefer to take the average of a few tests but I don't have the time to do it (is there a tool for it?).