@nolanlawson 2017-06-22
> won't it just increase latency on poor connections due to the extra round trip requests?

That can be fixed by adding `<link rel=preload>`s for each of the chunks that we expect are very likely to get loaded by the user. I think this would be a smart addition; @sorin-davidoi you can add it to `application.html.haml` like so:

```haml
%link{:rel => "preload", :as => "script", :src => (asset_pack_path 'application.js')}/
```

Also for your Lighthouse score, what page did you test against? There are a few I think we want to prioritize:

* `/web/getting-started` – main home
* `/web/timelines/home` – mobile PWA home (defined in `manifest.json`)

I also imagine there are people who bookmark local/federated/home timelines, so those would be worthwhile to preload as well.

Overall, I believe it is worthwhile to only load those routes that the user explicitly requested, since it limits overhead of JS parse/exec (and also memory). But avoiding extra network roundtrips is important, and also I'd like to see some deduplication of the shared modules across async chunks.