@sorin-davidoi 2017-06-22
I tested the local timeline (mobile UI). `<link rel=preload>` sounds nice, will try to add them! :+1: