@sorin-davidoi 2017-06-22
I think we could also lazy-load some reducers (see https://stackoverflow.com/questions/32968016/how-to-dynamically-load-reducers-for-code-splitting-in-a-redux-application), but that is a can of worms I don't want to open for now.