@beatrix-bitrot 2017-06-22
Is this ready for review?