@sorin-davidoi 2017-06-22
I'd say this is ready for review. Sometimes there is a crash when loading the HomeTimeline (hard to reproduce) which I need to look into, but otherwise, it looks alright.