@beatrix-bitrot 2017-06-23
nice, because it's a big PR and not my area of expertise I will test this branch on glitch.social to make my review