@nolanlawson 2017-06-23
In terms of bundle sizes this also looks great. I rebuilt and put a Webpack Bundle Analyzer view [here](http://bl.ocks.org/nolanlawson/raw/c8884382fb17c4cb5e733b8190e900c9/) for those who want to see. (Protip: make a Gist with an index.html and bl.ocks.org can host it for you. :smiley:)