@nolanlawson 2017-06-29
BTW http2 is recommended [in the tootsuite/documentation nginx config](https://github.com/tootsuite/documentation/blob/master/Running-Mastodon/Production-guide.md#nginx) so hopefully most admins already have it.