@unarist 2017-06-30
Component which is loaded before also seem to be lazy loaded.

![image](https://user-images.githubusercontent.com/705555/27738094-670d380e-5de5-11e7-8188-a57252f37bac.png)

I think it can be avoided by caching result of `import()`. (sample: https://github.com/unarist/mastodon/commit/4cf1fde8a2834cc42a2daa46da4d4d3270fef1ba)

![image](https://user-images.githubusercontent.com/705555/27739134-8a011a76-5de8-11e7-84fa-dc52b7f14c30.png)

Pros: avoid flashing of loading view, slightly reduces total time
Cons: single script execution will be longer, i.e. blocks browser longer


How about this?