@nolanlawson 2017-06-30
Yup sounds like a good tradeoff!