@nolanlawson 2017-07-04
@Gargron I propose we merge this PR along with https://github.com/unarist/mastodon/commit/4cf1fde8a2834cc42a2daa46da4d4d3270fef1ba; it should be a big win for the initial bundle size and I also really like how it cleans up the async route logic. :)