@sorin-davidoi 2017-07-08
@Gargron Open the Network tab in the Chrome development tools and tick "Disable cache".

For columns:
 - Loading state - Load a timeline, throttle the connection to GPRS, click on a toot.
 - Error state - Load a timeline, tick the "Offline" checkbox, click on a toot. After the error state is shown you can uncheck "Offline" and click retry, the normal page should appear.

For modals:
 - Loading state - throttle the connection to GPRS and open a media modal
 - Error state - tick "Offline" and open a media modal - same behavior with retry